Tax calculation correction. Updated README.
Change was to add a new rounded variable named itemTaxRounded after calculating the item's total tax. This new rounded variable was set to the return value of the roundTo5() with itemTaxNoRound as a parameter. Then this rounded variable was used in calculating the pricePlusTaxes for the item. It was also added per item (+=) to the totalTax variable. Also, updating the README.